Pull out shader source into new module
As we add more logic to the shader sources, it will be easier to reason
about each piece. Of note, the shader module really only contains the
generic gl logic, which seems fine.
