+++ lib/graphics/src/shader.rs
@@ -8,6 +8,7 @@ use std::ffi::CString;
 use gl::types::*;
 
 use vertex;
+use shader_source;
 
 pub struct Program {
     addr: GLuint,
@@ -88,7 +89,7 @@ pub struct Shader {
 }
 
 impl Shader {
-    pub fn new(shader_source: ShaderSource, shader_ty: GLShaderEnum) -> Shader {
+    pub fn new(shader_source: shader_source::ShaderSource, shader_ty: GLShaderEnum) -> Shader {
         let src = shader_source.source_glsl;
         let ty = shader_ty.to_glenum();
         let shader;
@@ -146,40 +147,3 @@ impl Shader {
         }
     }
 }
-
-pub struct ShaderSource {
-    pub source_glsl: &'static str,
-    pub var_name: &'static str,
-    pub stride: gl::types::GLsizei,
-    pub offset: usize,
-}
-
-const VS_SRC: &'static str = r#"#version 150
-    in vec2 position;
-    in vec3 color;
-    out vec3 attr_color;
-    void main() {
-       attr_color = color;
-       gl_Position = vec4(position, 0.0, 1.0);
-    }"#;
-
-pub const COLOR_VERTEX_SOURCE: ShaderSource = ShaderSource {
-    source_glsl: VS_SRC,
-    var_name: "position",
-    stride: 2,
-    offset: 0,
-};
-
-const FS_SRC: &'static str = r#"#version 150
-    in vec3 attr_color;
-    out vec4 out_color;
-    void main() {
-       out_color = vec4(attr_color, 1.0);
-    }"#;
-
-pub const COLOR_FRAGMENT_SOURCE: ShaderSource = ShaderSource {
-    source_glsl: FS_SRC,
-    var_name: "color",
-    stride: 3,
-    offset: 2,
-};
